all: add node.go
