+++ stores/hostdb_test.go
@@ -12,8 +12,8 @@ import (
 	rhpv2 "go.sia.tech/core/rhp/v2"
 	rhpv3 "go.sia.tech/core/rhp/v3"
 	"go.sia.tech/core/types"
+	"go.sia.tech/coreutils/chain"
 	"go.sia.tech/renterd/api"
-	"go.sia.tech/renterd/internal/chain"
 	sql "go.sia.tech/renterd/stores/sql"
 )
